+++ src/collective/contact/facetednav/browser/static/javascript.js
@@ -45,7 +45,7 @@ contactfacetednav.init = function() {
             closeselector: '[name="form.buttons.cancel"]',
             noform: function(el, pbo) {
                 Faceted.Form.do_form_query();
-                return 'close';},
+                return 'close';}
 
         });
         jQuery('#faceted-add a.faceted-add-contact').prepOverlay({
@@ -53,9 +53,9 @@ contactfacetednav.init = function() {
             filter: common_content_filter,
             formselector: '#oform',
             closeselector: '[name="oform.buttons.cancel"]',
-            noform: function(el, pbo) {
+            noform: function() {
                 Faceted.Form.do_form_query();
-                return 'close';},
+                return 'close';}
         });
     });
 };
@@ -76,7 +76,7 @@ contactfacetednav.Contact = Backbone.Model.extend({
     defaults : {
         id : "?????????????????",
         path : "?????????????????",
-        selected : false,
+        selected : false
     },
 
     initialize : function Contact() {
@@ -101,7 +101,6 @@ contactfacetednav.Contact = Backbone.Model.extend({
         var input = jQuery('#contact-' + this.get('id'));
         if (input.length === 0) {
             // Not in current page
-            return;
         }
         else if (this.isSelected()) {
             input.prop('checked', true);
@@ -120,7 +119,7 @@ contactfacetednav.Contacts = Backbone.Collection.extend({
     initialize : function() {
         this.bind('change', this.render);
         this.fetch({
-            success : function(model, response) {
+            success : function(model) {
                 model.each(function(contact) {
                     contact.render();
                 });
@@ -132,16 +131,20 @@ contactfacetednav.Contacts = Backbone.Collection.extend({
         return baseURL + '/json-contacts?' + jQuery.param(Faceted.SortedQuery());
     },
     selectAll: function(){
+        this.unbind('change', this.render);
         this.each(function(contact){
             contact.setSelected(true);
         });
-        this.trigger('select-all');
+        this.bind('change', this.render);
+        this.trigger('change');
     },
     unselectAll: function(){
+        this.unbind('change', this.render);
         this.each(function(contact){
             contact.setSelected(false);
         });
-        this.trigger('unselect-all');
+        this.bind('change', this.render);
+        this.trigger('change');
     },
     selection: function(){
         return this.filter(function(contact){
@@ -231,7 +234,7 @@ contactfacetednav.delete_selection = function(confirm_msg){
                         right: right + 'px',
                         top: '200px',
                         width: width + 'px',
-                        zIndex: 9999,
+                        zIndex: 9999
                     });
                     var form = divOverlay.find('form');
 
@@ -279,9 +282,9 @@ jQuery(Faceted.Events).bind(Faceted.Events.AJAX_QUERY_SUCCESS, function(){
                 filter: common_content_filter,
                 formselector: '#delete_confirmation',
                 closeselector: '[name="cancel"]',
-                noform: function(el, pbo) {
+                noform: function() {
                     Faceted.Form.do_form_query();
-                    return 'close';},
+                    return 'close';}
               });
     });
     jQuery('#faceted-results .edit-contact').each(function(){
@@ -291,9 +294,9 @@ jQuery(Faceted.Events).bind(Faceted.Events.AJAX_QUERY_SUCCESS, function(){
                 filter: common_content_filter,
                 formselector: '#form',
                 closeselector: '[name="form.button.cancel"]',
-                noform: function(el, pbo) {
+                noform: function() {
                     Faceted.Form.do_form_query();
-                    return 'close';},
+                    return 'close';}
               });
     });
 });